✨ feat: X축 라벨 설정 #123
