+++ FE/src/utils/drawUpperYAxis.ts
@@ -1,0 +1,48 @@
+import { Padding, StockChartUnit } from '../types.ts';
+import { makeYLabels } from './makeLabels.ts';
+
+export const drawUpperYAxis = (
+  ctx: CanvasRenderingContext2D,
+  data: StockChartUnit[],
+  width: number,
+  height: number,
+  padding: Padding,
+  weight: number = 0,
+) => {
+  const values = data
+    .map((d) => [+d.stck_hgpr, +d.stck_lwpr, +d.stck_clpr, +d.stck_oprc])
+    .flat();
+  const yMax = Math.round(Math.max(...values) * (1 + weight));
+  const yMin = Math.round(Math.min(...values) * (1 - weight));
+
+  ctx.clearRect(
+    0,
+    0,
+    width + padding.left + padding.right,
+    height + padding.top + padding.bottom,
+  );
+
+  const labels = makeYLabels(yMax, yMin, 3);
+
+  ctx.font = '24px sans-serif';
+  ctx.fillStyle = '#000';
+  ctx.textAlign = 'left';
+  ctx.textBaseline = 'middle';
+  ctx.beginPath();
+
+  labels.forEach((label) => {
+    const yPos =
+      padding.top + height - ((label - yMin) / (yMax - yMin)) * height;
+
+    const formattedValue = label.toLocaleString();
+    ctx.moveTo(0, yPos);
+    ctx.lineTo(padding.left, yPos);
+    ctx.fillText(formattedValue, width / 2 + padding.left, yPos);
+  });
+
+  ctx.moveTo(padding.left, 0);
+  ctx.lineTo(padding.left, height + padding.top + padding.bottom);
+  ctx.strokeStyle = '#D2DAE0';
+  ctx.lineWidth = 2;
+  ctx.stroke();
+};
